[DFI-533] vm mod: module minor refactoring: function names and godocs, files rearranging
