+++ x/vm/handler.go
@@ -5,19 +5,19 @@ import (
 	sdkErrors "github.com/cosmos/cosmos-sdk/types/errors"
 )
 
-// New message handler for PoA module.
-func NewHandler(keeper Keeper) sdk.Handler {
+// NewHandler creates sdk.Msg type messages handler.
+func NewHandler(k Keeper) sdk.Handler {
 	return func(ctx sdk.Context, msg sdk.Msg) (*sdk.Result, error) {
-		// settings actual context for ds.
+		// setup current (actual) DS context (that is also done in the BeginBlock
 		// TODO: move it to base app and set before transaction execution maybe? or find way to have actual context always
-		keeper.SetDSContext(ctx)
+		k.SetDSContext(ctx)
 
 		switch msg := msg.(type) {
 		case MsgDeployModule:
-			return handleMsgDeploy(ctx, keeper, msg)
+			return handleMsgDeploy(ctx, k, msg)
 
 		case MsgExecuteScript:
-			return handleMsgScript(ctx, keeper, msg)
+			return handleMsgScript(ctx, k, msg)
 
 		default:
 			return nil, sdkErrors.Wrapf(sdkErrors.ErrUnknownRequest, "unrecognized vm msg type: %v", msg.Type())
@@ -25,16 +25,18 @@ func NewHandler(keeper Keeper) sdk.Handler {
 	}
 }
 
-func handleMsgScript(ctx sdk.Context, keeper Keeper, msg MsgExecuteScript) (*sdk.Result, error) {
-	if err := keeper.ExecuteScript(ctx, msg); err != nil {
+// handleMsgScript handles MsgExecuteScript message.
+func handleMsgScript(ctx sdk.Context, k Keeper, msg MsgExecuteScript) (*sdk.Result, error) {
+	if err := k.ExecuteScript(ctx, msg); err != nil {
 		return nil, err
 	}
 
 	return &sdk.Result{Events: ctx.EventManager().Events()}, nil
 }
 
-func handleMsgDeploy(ctx sdk.Context, keeper Keeper, msg MsgDeployModule) (*sdk.Result, error) {
-	if err := keeper.DeployContract(ctx, msg); err != nil {
+// handleMsgDeploy handles MsgDeployModule message.
+func handleMsgDeploy(ctx sdk.Context, k Keeper, msg MsgDeployModule) (*sdk.Result, error) {
+	if err := k.DeployContract(ctx, msg); err != nil {
 		return nil, err
 	}
