+++ x/vm/internal/keeper/keeper_gov.go
@@ -8,8 +8,8 @@ import (
 )
 
 // ScheduleProposal checks if proposal can be added to gov proposal queue and adds it.
-func (keeper Keeper) ScheduleProposal(ctx sdk.Context, pProposal types.PlannedProposal) error {
-	keeper.modulePerms.AutoCheck(types.PermInit)
+func (k Keeper) ScheduleProposal(ctx sdk.Context, pProposal types.PlannedProposal) error {
+	k.modulePerms.AutoCheck(types.PermInit)
 
 	if pProposal.GetPlan().Height <= ctx.BlockHeight() {
 		return sdkErrors.Wrapf(
@@ -18,32 +18,32 @@ func (keeper Keeper) ScheduleProposal(ctx sdk.Context, pProposal types.PlannedPr
 			pProposal.GetPlan().Height, ctx.BlockHeight())
 	}
 
-	keeper.addProposalToQueue(ctx, pProposal)
+	k.addProposalToQueue(ctx, pProposal)
 
 	return nil
 }
 
 // RemoveProposalFromQueue removes proposal from the gov proposal queue.
-func (keeper Keeper) RemoveProposalFromQueue(ctx sdk.Context, id uint64) {
-	keeper.modulePerms.AutoCheck(types.PermInit)
+func (k Keeper) RemoveProposalFromQueue(ctx sdk.Context, id uint64) {
+	k.modulePerms.AutoCheck(types.PermInit)
 
 	queueKey := types.GetProposalQueueKey(id)
 
-	store := ctx.KVStore(keeper.storeKey)
+	store := ctx.KVStore(k.storeKey)
 	store.Delete(queueKey)
 }
 
 // IterateProposalsQueue iterates over gov proposal queue.
-func (keeper Keeper) IterateProposalsQueue(ctx sdk.Context, handler func(id uint64, pProposal types.PlannedProposal)) {
-	keeper.modulePerms.AutoCheck(types.PermInit)
+func (k Keeper) IterateProposalsQueue(ctx sdk.Context, handler func(id uint64, pProposal types.PlannedProposal)) {
+	k.modulePerms.AutoCheck(types.PermInit)
 
-	iterator := keeper.proposalQueueIterator(ctx)
+	iterator := k.proposalQueueIterator(ctx)
 	defer iterator.Close()
 
 	for ; iterator.Valid(); iterator.Next() {
 		id := types.SplitProposalQueueKey(iterator.Key())
 
-		p, err := keeper.unmarshalPlannedProposal(iterator.Value())
+		p, err := k.unmarshalPlannedProposal(iterator.Value())
 		if err != nil {
 			panic(err)
 		}
@@ -53,14 +53,14 @@ func (keeper Keeper) IterateProposalsQueue(ctx sdk.Context, handler func(id uint
 }
 
 // getNextProposalID returns next gov proposal queue ID.
-func (keeper Keeper) getNextProposalID(ctx sdk.Context) (id uint64) {
-	store := ctx.KVStore(keeper.storeKey)
+func (k Keeper) getNextProposalID(ctx sdk.Context) (id uint64) {
+	store := ctx.KVStore(k.storeKey)
 	if !store.Has(types.ProposalIDKey) {
 		return 0
 	}
 
 	bz := store.Get(types.ProposalIDKey)
-	if err := keeper.cdc.UnmarshalBinaryLengthPrefixed(bz, &id); err != nil {
+	if err := k.cdc.UnmarshalBinaryLengthPrefixed(bz, &id); err != nil {
 		panic(err)
 	}
 
@@ -68,39 +68,39 @@ func (keeper Keeper) getNextProposalID(ctx sdk.Context) (id uint64) {
 }
 
 // setProposalID updates gov proposal queue last ID.
-func (keeper Keeper) setProposalID(ctx sdk.Context, id uint64) {
-	store := ctx.KVStore(keeper.storeKey)
+func (k Keeper) setProposalID(ctx sdk.Context, id uint64) {
+	store := ctx.KVStore(k.storeKey)
 
-	bz := keeper.cdc.MustMarshalBinaryLengthPrefixed(id)
+	bz := k.cdc.MustMarshalBinaryLengthPrefixed(id)
 	store.Set(types.ProposalIDKey, bz)
 }
 
 // addProposalToQueue adds proposal to the gov proposal queue.
-func (keeper Keeper) addProposalToQueue(ctx sdk.Context, pProposal types.PlannedProposal) {
-	id := keeper.getNextProposalID(ctx)
+func (k Keeper) addProposalToQueue(ctx sdk.Context, pProposal types.PlannedProposal) {
+	id := k.getNextProposalID(ctx)
 	queueKey := types.GetProposalQueueKey(id)
 
-	store := ctx.KVStore(keeper.storeKey)
-	bz := keeper.cdc.MustMarshalBinaryLengthPrefixed(pProposal)
+	store := ctx.KVStore(k.storeKey)
+	bz := k.cdc.MustMarshalBinaryLengthPrefixed(pProposal)
 	store.Set(queueKey, bz)
-	keeper.setProposalID(ctx, id)
+	k.setProposalID(ctx, id)
 }
 
 // proposalQueueIterator returns gov proposal queue iterator.
-func (keeper Keeper) proposalQueueIterator(ctx sdk.Context) sdk.Iterator {
-	store := ctx.KVStore(keeper.storeKey)
+func (k Keeper) proposalQueueIterator(ctx sdk.Context) sdk.Iterator {
+	store := ctx.KVStore(k.storeKey)
 	return store.Iterator(types.ProposalQueuePrefix, sdk.PrefixEndBytes(types.ProposalQueuePrefix))
 }
 
 // unmarshalPlannedProposal unmarshals stored PlannedProposal to a known concrete type.
-func (keeper Keeper) unmarshalPlannedProposal(bz []byte) (types.PlannedProposal, error) {
+func (k Keeper) unmarshalPlannedProposal(bz []byte) (types.PlannedProposal, error) {
 	pStdlib := types.StdlibUpdateProposal{}
-	if err := keeper.cdc.UnmarshalBinaryLengthPrefixed(bz, &pStdlib); err == nil {
+	if err := k.cdc.UnmarshalBinaryLengthPrefixed(bz, &pStdlib); err == nil {
 		return pStdlib, nil
 	}
 
 	pTest := types.TestProposal{}
-	if err := keeper.cdc.UnmarshalBinaryLengthPrefixed(bz, &pTest); err == nil {
+	if err := k.cdc.UnmarshalBinaryLengthPrefixed(bz, &pTest); err == nil {
 		return pTest, nil
 	}
