+++ x/vm/internal/types/default.go
@@ -1,0 +1,19 @@
+package types
+
+const (
+	ModuleName   = "vm"
+	StoreKey     = ModuleName
+	RouterKey    = ModuleName
+	GovRouterKey = ModuleName
+	//
+	VmGasPrice       = 1 // gas unit price for VM execution
+	VmUnknownTagType = -1
+	// VM Event to sdk.Event conversion params
+	EventTypeProcessingGas = 10000 // initial gas for processing event type.
+	EventTypeNoGasLevels   = 2     // defines number of nesting levels that do not charge gas
+)
+
+var (
+	KeyDelimiter = []byte(":")
+	KeyGenesis   = []byte("gen")
+)
