+++ x/vm/internal/types/msg_execute_script_test.go
@@ -3,7 +3,6 @@
 package types
 
 import (
-	"encoding/json"
 	"testing"
 
 	sdk "github.com/cosmos/cosmos-sdk/types"
@@ -15,40 +14,6 @@ import (
 	"github.com/dfinance/dnode/helpers/tests/utils"
 	"github.com/dfinance/dnode/x/common_vm"
 )
-
-func getMsgSignBytes(t *testing.T, msg sdk.Msg) []byte {
-	bc, err := json.Marshal(msg)
-	if err != nil {
-		t.Fatal(err)
-	}
-
-	return sdk.MustSortJSON(bc)
-}
-
-// Test MsgDeployModule.
-func TestMsgDeployModule(t *testing.T) {
-	t.Parallel()
-
-	acc := sdk.AccAddress([]byte("addr1"))
-	code := make(Contract, 128)
-	msg := NewMsgDeployModule(acc, code)
-
-	require.Equal(t, msg.Signer, acc)
-	require.Equal(t, msg.Module, code)
-	require.NoError(t, msg.ValidateBasic())
-	require.Equal(t, RouterKey, msg.Route())
-	require.Equal(t, MsgDeployModuleType, msg.Type())
-	require.Equal(t, msg.GetSigners(), []sdk.AccAddress{acc})
-	require.Equal(t, getMsgSignBytes(t, msg), msg.GetSignBytes())
-
-	msg = NewMsgDeployModule([]byte{}, code)
-	require.Empty(t, msg.Signer)
-	utils.CheckExpectedErr(t, sdkErrors.ErrInvalidAddress, msg.ValidateBasic())
-
-	msg = NewMsgDeployModule(acc, Contract{})
-	require.Empty(t, msg.Module)
-	utils.CheckExpectedErr(t, ErrEmptyContract, msg.ValidateBasic())
-}
 
 // Test MsgExecuteScript.
 func TestMsgExecuteScript(t *testing.T) {
